First attempt at moving download protocol into Transceiver (not quite working)
