+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/protocols/download/GatherDownloadConnection.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.manchester.spinnaker.front_end.download;
+package uk.ac.manchester.spinnaker.protocols.download;
 
 import static java.lang.System.nanoTime;
 import static java.nio.ByteBuffer.allocate;
