+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/protocols/download/GatherProtocolMessage.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.manchester.spinnaker.front_end.download;
+package uk.ac.manchester.spinnaker.protocols.download;
 
 import static uk.ac.manchester.spinnaker.messages.sdp.SDPHeader.Flag.REPLY_NOT_EXPECTED;
