+++ SpiNNaker-front-end/src/main/java/uk/ac/manchester/spinnaker/front_end/download/DataGatherer.java
@@ -16,25 +16,16 @@
 package uk.ac.manchester.spinnaker.front_end.download;
 
 import static difflib.DiffUtils.diff;
-import static java.lang.String.format;
 import static java.lang.System.getProperty;
-import static java.lang.Thread.sleep;
 import static java.nio.ByteBuffer.allocate;
 import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toUnmodifiableList;
 import static org.slf4j.LoggerFactory.getLogger;
 import static uk.ac.manchester.spinnaker.front_end.Constants.PARALLEL_SIZE;
-import static uk.ac.manchester.spinnaker.front_end.download.MissingSequenceNumbersMessage.createMessages;
-import static uk.ac.manchester.spinnaker.messages.Constants.SDP_PAYLOAD_WORDS;
-import static uk.ac.manchester.spinnaker.messages.Constants.WORD_SIZE;
 import static uk.ac.manchester.spinnaker.utils.ByteBufferUtils.sliceUp;
-import static uk.ac.manchester.spinnaker.utils.MathUtils.ceildiv;
 
 import java.io.IOException;
-import java.nio.BufferOverflowException;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
-import java.util.BitSet;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -82,45 +73,6 @@ public abstract class DataGatherer extends BoardLocalSupport
 	 * Logger for the gatherer.
 	 */
 	protected static final Logger log = getLogger(DataGatherer.class);
-
-	/** The maximum number of times to retry. */
-	private static final int TIMEOUT_RETRY_LIMIT = 15;
-
-	/**
-	 * The time delay between sending each message. In
-	 * {@linkplain java.util.concurrent.TimeUnit#MILLISECONDS milliseconds}.
-	 */
-	private static final int DELAY_PER_SEND = 10;
-
-	/**
-	 * The timeout when receiving a message. In
-	 * {@linkplain java.util.concurrent.TimeUnit#MILLISECONDS milliseconds}.
-	 */
-	private static final int TIMEOUT_PER_RECEIVE = 2000;
-
-	/**
-	 * The <i>extra</i> timeout for processing the message queue. In
-	 * {@linkplain java.util.concurrent.TimeUnit#MILLISECONDS milliseconds}.
-	 */
-	private static final int INTERNAL_DELAY = 100;
-
-	/**
-	 * What is the maximum number of payload <em>words</em> in a packet that
-	 * also has a sequence number? This is one less than the total maximum
-	 * number of words in an SDP packet; that extra word is the control word
-	 * which encodes the sequence number and the end-of-stream flag.
-	 */
-	private static final int DATA_WORDS_PER_PACKET = SDP_PAYLOAD_WORDS - 2;
-
-	/**
-	 * Mask used to pick out the bit that says whether a sequence number is the
-	 * last in a stream.
-	 */
-	private static final int LAST_MESSAGE_FLAG_BIT_MASK = 0x80000000;
-
-	/** Message used to report problems. */
-	private static final String TIMEOUT_MESSAGE = "failed to hear from the "
-			+ "machine (please try removing firewalls)";
 
 	private static final String SPINNAKER_COMPARE_DOWNLOAD =
 			getProperty("spinnaker.compare.download");
@@ -250,7 +202,6 @@ public abstract class DataGatherer extends BoardLocalSupport
 		for (var g : gatherers) {
 			var workitems = new ArrayList<WorkItems>();
 			for (var m : g.getMonitors()) {
-				m.updateTransactionIdFromMachine(txrx);
 
 				for (var p : m.getPlacements()) {
 					var regions = new ArrayList<Region>();
@@ -325,15 +276,10 @@ public abstract class DataGatherer extends BoardLocalSupport
 	 */
 	private void fastDownload(List<WorkItems> work,
 			ChipLocation gathererChip, IPTag ipTag)
-			throws IOException, StorageException, TimeoutException,
+			throws IOException, StorageException,
 			ProcessException, InterruptedException {
-		try (var scpConn = txrx.createScpConnection(
-					gathererChip, ipTag.getBoardAddress());
-				var conn = new GatherDownloadConnection(scpConn);
-				var c = new BoardLocal(conn.getChip())) {
-			conn.setIPTag(txrx, ipTag);
-			log.info("processing fast downloads for {}", conn.getChip());
-			var dl = new Downloader(conn);
+		try (var c = new BoardLocal(gathererChip)) {
+			log.info("processing fast downloads for {}", gathererChip);
 			for (var item : work) {
 				for (var region : item.regions) {
 					/*
@@ -341,7 +287,8 @@ public abstract class DataGatherer extends BoardLocalSupport
 					 * retrieves for that recording region have to be done the
 					 * same way to get the data in the DB in the right order.
 					 */
-					var data = dl.doDownload(item.monitor, region);
+					var data = txrx.readMemoryFast(gathererChip, ipTag,
+							item.monitor, region.startAddress, region.size);
 					if (SPINNAKER_COMPARE_DOWNLOAD != null) {
 						compareDownloadWithSCP(region, data);
 					}
@@ -470,336 +417,6 @@ public abstract class DataGatherer extends BoardLocalSupport
 	protected abstract void storeData(Region r, ByteBuffer data)
 			throws StorageException;
 
-	/**
-	 * Class used to manage a download. Every instance <em>must only</em> ever
-	 * be used from one thread.
-	 *
-	 * @author Donal Fellows
-	 */
-	private final class Downloader {
-		private final GatherDownloadConnection conn;
 
-		/**
-		 * Whether a packet has previously been received on this connection
-		 * since it was configured to talk to the current core; if not, it's
-		 * probably a dead connection or problem with the core causing the
-		 * failure.
-		 */
-		private boolean received;
-
-		private int timeoutcount;
-
-		/** A flag for each packet sequence number that is expected. */
-		private BitSet expectedSeqNums;
-
-		private int maxSeqNum;
-
-		private ByteBuffer dataReceiver;
-
-		private Monitor monitorCore;
-
-		private List<Integer> lastRequested;
-
-		/**
-		 * Create an instance.
-		 *
-		 * @param connection
-		 *            The connection used to send messages.
-		 */
-		private Downloader(GatherDownloadConnection connection) {
-			conn = connection;
-		}
-
-		/**
-		 * Do the downloading.
-		 *
-		 * @param extraMonitor
-		 *            Where to download from.
-		 * @param region
-		 *            What to download.
-		 * @return The downloaded data, or {@code null} if an unrecoverable
-		 *         error occurred.
-		 * @throws IOException
-		 *             If anything unexpected goes wrong.
-		 * @throws TimeoutException
-		 *             If a download times out unrecoverably.
-		 * @throws ProcessException
-		 *             If anything unexpected goes wrong.
-		 * @throws InterruptedException
-		 *             If communications are interrupted.
-		 */
-		ByteBuffer doDownload(Monitor extraMonitor, Region region)
-				throws IOException, TimeoutException, ProcessException,
-				InterruptedException {
-			monitorCore = extraMonitor;
-			dataReceiver = allocate(region.size);
-			/*
-			 * Tricky point: where an amount of data to be downloaded is exactly
-			 * a multiple of the number of payload words per packet, we need an
-			 * extra packet because the data gatherer executable sends the data
-			 * before it knows whether it has reached the end (and the extra
-			 * monitors don't know about the chunking).
-			 *
-			 * This translates into needing to add one here.
-			 */
-			maxSeqNum = ceildiv(region.size, DATA_WORDS_PER_PACKET * WORD_SIZE);
-			expectedSeqNums = new BitSet(maxSeqNum);
-			expectedSeqNums.set(0, maxSeqNum);
-			lastRequested = expectedSeqs();
-			received = false;
-			timeoutcount = 0;
-			monitorCore.updateTransactionId();
-			log.debug(
-					"extracting data from {} with size {} with "
-							+ "transaction id {}",
-					region.startAddress, region.size,
-					monitorCore.getTransactionId());
-			conn.sendStart(monitorCore.asCoreLocation(), region.startAddress,
-					region.size, monitorCore.getTransactionId());
-			try {
-				boolean finished;
-				do {
-					finished = processOnePacket(TIMEOUT_PER_RECEIVE,
-							monitorCore.getTransactionId());
-				} while (!finished);
-				conn.sendClear(monitorCore.asCoreLocation(),
-						monitorCore.getTransactionId());
-			} catch (TimeoutException e) {
-				if (received) {
-					log.warn(
-							"received only some of the packets from <{}> "
-									+ "for {}; has something crashed?",
-							monitorCore, region);
-				}
-				throw e;
-			} finally {
-				if (!received) {
-					log.warn("never actually received any packets from "
-							+ "<{}> for {}", monitorCore, region);
-				}
-			}
-			dataReceiver.position(0);
-			try {
-				return dataReceiver;
-			} finally {
-				dataReceiver = null;
-			}
-		}
-
-		/**
-		 * Take one message off the queue and process it.
-		 *
-		 * @param timeout
-		 *            How long to wait for the queue to deliver a packet, in
-		 *            milliseconds.
-		 * @param transactionId
-		 *            The transaction id of this stream.
-		 * @return True if we have finished.
-		 * @throws IOException
-		 *             If packet reception or retransmission requesting fails.
-		 * @throws TimeoutException
-		 *             If we have a full timeout, or if we are flailing around,
-		 *             making no progress.
-		 * @throws InterruptedException
-		 *             If communications are interrupted.
-		 */
-		private boolean processOnePacket(int timeout, int transactionId)
-				throws IOException, TimeoutException, InterruptedException {
-			var p = conn.getNextPacket(timeout + INTERNAL_DELAY);
-			if (p.hasRemaining()) {
-				received = true;
-				timeoutcount = 0;
-				return processData(p, transactionId);
-			}
-			log.debug("failed to receive on connection {}.", conn);
-			return processTimeout(transactionId);
-		}
-
-		/**
-		 * Process a single received packet.
-		 *
-		 * @param data
-		 *            The content of the packet.
-		 * @param transactionId
-		 *            The transaction id of this stream.
-		 * @return True if we have finished.
-		 * @throws IOException
-		 *             If the packet is an end-of-stream packet yet there are
-		 *             packets outstanding, and the retransmission causes an
-		 *             error.
-		 * @throws TimeoutException
-		 *             If we are flailing around, making no progress.
-		 * @throws InterruptedException
-		 *             If we are interrupted.
-		 */
-		private boolean processData(ByteBuffer data, int transactionId)
-				throws IOException, TimeoutException, InterruptedException {
-			int seqNum = data.getInt();
-			int responseTransactionId = data.getInt();
-
-			if (responseTransactionId != transactionId) {
-				return false;
-			}
-
-			var isEndOfStream = ((seqNum & LAST_MESSAGE_FLAG_BIT_MASK) != 0);
-			seqNum &= ~LAST_MESSAGE_FLAG_BIT_MASK;
-
-			if (seqNum > maxSeqNum || seqNum < 0) {
-				throw new InsaneSequenceNumberException(maxSeqNum, seqNum);
-			}
-			int len = data.remaining();
-			if (len != DATA_WORDS_PER_PACKET * WORD_SIZE && len != 0
-					&& seqNum < maxSeqNum - 1) {
-				log.warn("short packet ({} bytes) in non-terminal position "
-						+ "(seq: {})", len, seqNum);
-			}
-			if (data.hasRemaining()) {
-				int offset = seqNum * DATA_WORDS_PER_PACKET * WORD_SIZE;
-				if (offset < 0) {
-					// Off the start!
-					throw new ReceivingBufferOverflowedException(
-							len, offset, dataReceiver.limit());
-				}
-				// The IllegalArgumentException on failure here is useful
-				dataReceiver.position(offset);
-				try {
-					dataReceiver.put(data);
-				} catch (BufferOverflowException ignored) {
-					// There's no info in that exception
-					throw new ReceivingBufferOverflowedException(
-							len, offset, dataReceiver.limit());
-				}
-				expectedSeqNums.clear(seqNum);
-			}
-			if (!isEndOfStream) {
-				return false;
-			}
-			return retransmitMissingSequences(transactionId);
-		}
-
-		/**
-		 * Process the fact that the message queue was in a timeout state.
-		 *
-		 * @param transactionId
-		 *            The transaction id of this stream
-		 * @return True if we have finished.
-		 * @throws IOException
-		 *             If there are packets outstanding, and the retransmission
-		 *             causes an error.
-		 * @throws TimeoutException
-		 *             If we have a full timeout.
-		 * @throws InterruptedException
-		 *             If we are interrupted.
-		 */
-		private boolean processTimeout(int transactionId)
-				throws IOException, TimeoutException, InterruptedException {
-			if (++timeoutcount > TIMEOUT_RETRY_LIMIT) {
-				log.error(TIMEOUT_MESSAGE);
-				throw new TimeoutException();
-			}
-
-			// retransmit missing packets
-			log.debug("doing reinjection");
-			return retransmitMissingSequences(transactionId);
-		}
-
-		/**
-		 * Request that the extra monitor core retransmit some packets. Does
-		 * nothing if there are no packets missing.
-		 *
-		 * @param transactionId
-		 *            The transaction id of this stream
-		 * @return Whether there were really any packets to retransmit.
-		 * @throws IOException
-		 *             If there are failures.
-		 * @throws TimeoutException
-		 *             If we are flailing around, making no progress.
-		 * @throws InterruptedException
-		 *             If interrupted.
-		 */
-		private boolean retransmitMissingSequences(int transactionId)
-				throws IOException, TimeoutException, InterruptedException {
-			int numMissing = expectedSeqNums.cardinality();
-			if (numMissing < 1) {
-				return true;
-			}
-			log.debug("there are {} missing packets in message from {}",
-					numMissing, monitorCore);
-
-			// Build a list of the sequence numbers of all missing packets
-			var missingSeqs = expectedSeqs();
-			missCount += numMissing;
-
-			log.debug("missing sequence numbers: {}", missingSeqs);
-			if (missingSeqs.size() > lastRequested.size()) {
-				log.warn("what is going on?");
-				log.warn("last:{}", lastRequested);
-				log.warn("this:{}", missingSeqs);
-			}
-			lastRequested = missingSeqs;
-
-			// Transmit missing sequences as a new SDP Packet
-			for (var msg : createMessages(monitorCore, missingSeqs,
-					transactionId)) {
-				sleep(DELAY_PER_SEND);
-				conn.sendMissing(msg);
-			}
-			return false;
-		}
-
-		/**
-		 * @return The expected sequence numbers, as an ordered list.
-		 */
-		private List<Integer> expectedSeqs() {
-			return expectedSeqNums.stream().boxed()
-					.collect(toUnmodifiableList());
-		}
-	}
 }
 
-/**
- * Exception that indicates a total (i.e., unrecoverable) failure to do a
- * download.
- *
- * @author Donal Fellows
- */
-final class TimeoutException extends Exception {
-	private static final long serialVersionUID = 1L;
-}
-
-/**
- * Exception that indicates a bad sequence number in a download.
- *
- * @author Donal Fellows
- */
-final class InsaneSequenceNumberException extends IllegalStateException {
-	private static final long serialVersionUID = 2L;
-
-	private static final String MID = "totally bad sequence";
-
-	private static final String END = "non-empty terminal-only packet";
-
-	InsaneSequenceNumberException(int maxNum, int seqNum) {
-		super(format("got insane sequence number %d: expected maximum %d (%s)",
-				seqNum, maxNum, (maxNum == seqNum ? END : MID)));
-	}
-}
-
-/**
- * Exception that indicates a download tried to write off the end of the
- * receiving buffer or otherwise misbehave. Don't use the standard exceptions
- * for this; they don't have any useful info at all.
- *
- * @see BufferOverflowException
- * @author Donal Fellows
- */
-final class ReceivingBufferOverflowedException extends RuntimeException {
-	private static final long serialVersionUID = 3L;
-
-	ReceivingBufferOverflowedException(int len, int start, int limit) {
-		super(format(
-				"failed to store %d bytes at position %d of %d "
-						+ "(overflow = %d bytes)",
-				len, start, limit, start + len - limit));
-	}
-}
